// src/app/pages/dashboard/dashboard-request/model/products-data.ts
export interface Products {
  product: string;
  serviceRequestNumber: number;
  serviceRequestDate: string;
  serviceRequestLine: number;
  typeId: number;
  businessId: number;
  businessName: string;
  contactId: number;
  contactName: string;
  contactPhoneNumber: string;
  contactEmail: string;
  paxCount: number;
  originDestinationLocation: string;
  dateOfJourneyCheckIn: string;
  classDays: string;
  createdUserName: string;
  createdUserEmail: string;
  createdUserPhone: string;
}